requirements.txt
alembic
annotated-types
anyio
bidict
blinker
boto3
botocore
certifi
charset-normalizer
click
contourpy
cycler
exceptiongroup
fastapi
filelock
Flask
Flask-Migrate
Flask-SocketIO
Flask-SQLAlchemy
fonttools
fsspec
greenlet
gunicorn
h11
idna
imageio
iniconfig
itsdangerous
Jinja2
jmespath
joblib
kiwisolver
lazy_loader
Mako
MarkupSafe
matplotlib
mpmath
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
opencv-python
packaging
pandas
pillow
pluggy
psutil
py-cpuinfo
pycryptodome
pydantic
pydantic-settings
pydantic_core
pyparsing
pytest
python-dateutil
python-dotenv
python-engineio
python-socketio
pytz
PyYAML
requests
s3transfer
scikit-image
scikit-learn
scipy
seaborn
simple-websocket
six
sniffio
SQLAlchemy
starlette
sympy
tabulate
threadpoolctl
tifffile
tomli
torch
torchvision
tqdm
triton
typing_extensions
tzdata
ultralytics
ultralytics-thop
unicorn
urllib3
uvicorn
Werkzeug
wsproto
